moved create_field_definition back to config_manager
